fix(processor): preserve queue helper test seams
